general info endpoint using JPQL

+++ backend/src/main/java/com/myback/dto/RegionToStatsDto.java
@@ -1,7 +1,5 @@
 package com.myback.dto;
 
-import java.math.BigDecimal;
-
 import lombok.*;
 
 @Getter
@@ -14,17 +12,18 @@ public class RegionToStatsDto {
     private String continent_name;
     private String region_name;
     private String country_name;
-    private int year;
-    private int population;
-    private BigDecimal gdp;
+    private Integer year;
+    private Long population;
+    private Long gdp;
 
-    // public RegionToStatsDto(String continent_name, String region_name, String country_name, int year, int population,
-    //         BigDecimal gdp) {
-    //     this.continent_name = continent_name;
-    //     this.region_name = region_name;
-    //     this.country_name = country_name;
-    //     this.year = year;
-    //     this.population = population;
-    //     this.gdp = gdp;
+
+    // public RegionToStatsDto(String cn, String rn, String con, int y, Integer p) {
+    //     this.continent_name = cn;
+    //     this.region_name = rn;
+    //     this.country_name = con;
+    //     this.year = y;
+    //     this.population = (null == p) ? 0 : p;
+    //     // this.gdp = g;
     // }
+
 }
